Changed data type of a cap bank's shorted groups to FLOAT instead of INT.

+++ Source/Libraries/openXDA.Model/TransmissionElements/CapBank.cs
@@ -187,7 +187,7 @@ namespace openXDA.Model
         /// <summary>
         /// Line 64
         /// </summary>
-        public int ShortedGroups { get; set; }
+        public double ShortedGroups { get; set; }
 
 
         /// <summary>
